reduce eth call footprint (#124)
* Remove execution channel (and always execute max attempt)

This is actually a rollback to previous semantics, where
just the max attempt is executed. We had the execution with
channel because we didn't want to wait to execute a obvious
opportunity. The reason for this is that the calculation
was slow before, but I intend to make it really fast with
this PR so I thought we could get rid of the complexity
and imprecision on execution.

* Move block_id into Block struct

* Add itertools crate

* Centralize reserves/balances fetching

* Remove reserves/balances fetching from execution